requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
Django==5.0.3
djangorestframework==3.14.0
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
yarl==1.9.4
